+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `failure_date` field to `ComputePlan` protocol buffer schema 
 - `IsComputePlanRunning` gRPC method in `ComputePlanService`
+- allow setting gRPC keepalive enforcement policy
 
 ### Removed